Added XID to logout URL
The link generated by {exp:freemember:logout_url} in EE 2.7+ fails because it does not include an XID. An XID is usually only required for POST submissions. However, since the act_logout method utilizes member_logout from the core EE member class, the secure_forms_check filter applies in this GET context. Not wanting to hack the core, I've added the XID hash to the logout URL that logout_url generates to bypass this issue.